test: Use net/http instead of martini.
The third party package does not make the code any shorter or easier to
understand. Most people are more familiar with net/http. It makes sense
to use it in this case.
Remove code that serves sourcemaps from GOROOT and GOPATH. I could not
verify it had any effect. Also, GOPATH variable may not be a valid path
if it contains more than one GOPATH workspace, so it's incorrect code
anyway. It can be re-added if needed.