Remove dependency on honnef.co/go/js/{dom,util}.
The dom package is large, so for users that don't already import it,
it will make this package much more lightweight. The approximate
amount of non-gziped, minified savings is on the order of 220 KB.

Not relying on dom package means having to repeat some of the work
it does (and increase the generated size for packages that _do_
import dom package, but by very little since there's not much overlap).
It's worth it since there's not much, and this library continues to
provide the same Go API as before.

Closes #15.
Updates #4.