Make Dial return net.Conn interface, not custom struct
* Make Dial return net.Conn interface, not custom struct.

This simplifies the documentation and package slightly.

Remove WriteString method. It's a wrapper around Send method, and is
not helpful in the context of a net.Conn, which deals with pure streams
of bytes. To send text frames, either use a WebSocket and Send, or use
type assertion to access the Send method from the returned net.Conn
implementation (it embeds a WebSocket).

Resolves #11.

* Minor style improvements, internal doc fixes.

Use less naked returns. They're harder to read.